requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.6
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.10.0
sqlparse==0.5.3
tzdata==2025.1
flake8
djangorestframework
markdown
django-filter
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
python-decouple
